requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-import-export==2.4.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
